requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
google-api-core==2.8.2
google-auth==2.9.0
google-cloud-bigquery~=3.4.0
google-cloud-core==2.3.1
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-status==1.47.0
idna==3.3
packaging==21.3
pip==22.1.2
proto-plus==1.20.6
protobuf>=4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
rsa==4.8
setuptools==63.1.0
six==1.16.0
typing_extensions==4.3.0
urllib3==1.26.10
wheel==0.37.1